requirements.txt
numpy>=1.24.0
scipy>=1.10.0
pyyaml>=6.0
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-timeout>=2.1.0
pytest-benchmark>=4.0.0
llama-cpp-python>=0.2.0
jsonschema>=4.0.0
aiofiles>=23.0.0
psutil>=5.0.0
flake8>=6.0.0
mypy>=1.0.0
bandit>=1.7.0
cryptography>=42.0.0
mujoco>=3.0.0
brian2>=2.5.0
tenseal>=0.3.0
